+++ R/finalize_grid.R
@@ -44,6 +44,10 @@ setMethod(
 
     x@table_string <- out
 
+    for (fn in x@lazy_finalize) {
+      x <- fn(x)
+    }
+
     return(x)
 })