+++ R/finalize_tabularray.R
@@ -2,6 +2,9 @@ setMethod(
   f = "finalize",
   signature = "tinytable_tabularray",
   definition = function(x, ...) {
+    for (fn in x@lazy_finalize) {
+      x <- fn(x)
+    }
     return(x)
   }
 )